Split select into base component
Using inheritance (shudder) to provide difference in functionality
between single and multi select
