+++ components/util/util.ts
@@ -14,7 +14,7 @@ export enum KeyCode {
 
 type RecursiveObject = { [name:string]:RecursiveObject };
 
-// This involves some fun type fuckery - this is essentially a function to retrive the value at a given path.
+// This involves some fun type fuckery (It can be likened to RAA) - this is essentially a function to retrive the value at a given path.
 // If anyone has a better way, please do let me know :)
 export function deepValue<T, U>(object:T, path:string):U {
     if (!object) {
