+++ demo/src/app/pages/test/test.page.html
@@ -10,8 +10,8 @@
         <template let-option #optionTemplate>
             <i class="child icon"></i>{{ option }}
         </template>
-        <new-sui-select [options]="options" [isSearchable]="true" [optionTemplate]="optionTemplate" #select>
-            <new-sui-select-option *ngFor="let option of select.availableOptions" [value]="option"></new-sui-select-option>
-        </new-sui-select>
+        <sui-select [options]="options" [isSearchable]="true" [optionTemplate]="optionTemplate" #select>
+            <sui-select-option *ngFor="let option of select.availableOptions" [value]="option"></sui-select-option>
+        </sui-select>
     </div>
 </demo-page-content>
